new folder structure, sand_tests WIP
