+++ test/landSaleWithReferral/subTests/fixtures.js
@@ -1,8 +1,8 @@
 // const {utils} = require("ethers");
 const {ethers, deployments, getNamedAccounts} = require("@nomiclabs/buidler");
-const MerkleTree = require("../../lib/merkleTree");
+const MerkleTree = require("../../../lib/merkleTree");
 const {getChainCurrentTime} = require("testUtils");
-const {createDataArray} = require("../../lib/merkleTreeHelper");
+const {createDataArray} = require("../../../lib/merkleTreeHelper");
 const {testLands, generateUserPermissions, setupUser} = require("./_testHelper");
 
 // Inputs
